// src/app/providers/models/formData.model.ts
export class FormData {
    mixnum: string = '';
    company: string = '';
    contractor: string = '';
    site: string = '';

    w1_unit: string = '';
    w2_unit: string = '';
    w3_unit: string = '';
    c1_unit: string = '';
    c2_unit: string = '';
    c3_unit: string = '';
    mad1_unit: string = '';
    mad2_unit: string = '';
    mad3_unit: string = ''
    w1_density: string = '';
    w2_density: string = '';
    w3_density: string = '';
    c1_density: string = '';
    c2_density: string = '';
    c3_density: string = '';
    mad1_density: string = '';
    mad2_density: string = '';
    mad3_density: string = '';

    s1_unit: string = '';
    s2_unit: string = '';
    s3_unit: string = '';
    g1_unit: string = '';
    g2_unit: string = '';
    g3_unit: string = '';
    ad1_unit: string = '';
    ad2_unit: string = '';
    ad3_unit: string = '';

    s1_density: string = '';
    s2_density: string = '';
    s3_density: string = '';
    g1_density: string = '';
    g2_density: string = '';
    g3_density: string = '';
    ad1_density: string = '';
    ad2_density: string = '';
    ad3_density: string = '';

    air: string = '';
    aggregate: string = '';
    volume: string = '';
    wet: string = '';


    clear() {
        this.mixnum = '';
        this.company = '';
        this.contractor = '';
        this.site = '';

        this.w1_unit = '';
        this.w2_unit = '';
        this.w3_unit = '';
        this.c1_unit = '';
        this.c2_unit = '';
        this.c3_unit = '';
        this.mad1_unit = '';
        this.mad2_unit = '';
        this.mad3_unit = '';

        this.w1_density = '';
        this.w2_density = '';
        this.w3_density = '';
        this.c1_density = '';
        this.c2_density = '';
        this.c3_density = '';
        this.mad1_density = '';
        this.mad2_density = '';
        this.mad3_density = '';
        
        this.s1_unit = '';
        this.s2_unit = '';
        this.s3_unit = '';
        this.g1_unit = '';
        this.g2_unit = '';
        this.g3_unit = '';
        this.ad1_unit = '';
        this.ad2_unit = '';
        this.ad3_unit = '';

        this.s1_density = '';
        this.s2_density = '';
        this.s3_density = '';
        this.g1_density = '';
        this.g2_density = '';
        this.g3_density = '';
        this.ad1_density = '';
        this.ad2_density = '';
        this.ad3_density = '';
        
        this.air = '';
        this.aggregate = '';
        this.volume = '';
        this.wet = '';
    }

}

export class Personal {
    mixnum: string = '';
    company: string = '';
    contractor: string = '';
    site: string = '';
}

export class Element {
    w1_unit: string = '';
    w2_unit: string = '';
    w3_unit: string = '';
    c1_unit: string = '';
    c2_unit: string = '';
    c3_unit: string = '';
    mad1_unit: string = '';
    mad2_unit: string = '';
    mad3_unit: string = '';

    w1_density: string = '';
    w2_density: string = '';
    w3_density: string = '';
    c1_density: string = '';
    c2_density: string = '';
    c3_density: string = '';
    mad1_density: string = '';
    mad2_density: string = '';
    mad3_density: string = '';
}

export class Element2 {
    s1_unit: string = '';
    s2_unit: string = '';
    s3_unit: string = '';
    g1_unit: string = '';
    g2_unit: string = '';
    g3_unit: string = '';
    ad1_unit: string = '';
    ad2_unit: string = '';
    ad3_unit: string = '';

    s1_density: string = '';
    s2_density: string = '';
    s3_density: string = '';
    g1_density: string = '';
    g2_density: string = '';
    g3_density: string = '';
    ad1_density: string = '';
    ad2_density: string = '';
    ad3_density: string = '';
}

export class Percentage {
    air: string = '';
    aggregate: string = '';
    volume: string = '';
    wet: string = '';
}